KAFKA-17469: Moved share external interfaces to share module.
