+++ build.gradle
@@ -1637,6 +1637,7 @@ project(':share-coordinator') {
     implementation project(':metadata')
     implementation project(':server')
     implementation project(':server-common')
+    implementation project(':share')
     implementation libs.slf4jApi
     implementation libs.metrics
