+++ checkstyle/import-control-share-coordinator.xml
@@ -56,7 +56,7 @@
             <allow pkg="org.apache.kafka.metadata" />
             <allow pkg="org.apache.kafka.server.common" />
             <allow pkg="org.apache.kafka.server.config" />
-            <allow pkg="org.apache.kafka.server.group.share" />
+            <allow pkg="org.apache.kafka.server.share" />
             <allow pkg="org.apache.kafka.server.record" />
             <allow pkg="org.apache.kafka.server.util" />
             <allow pkg="org.apache.kafka.server.util.timer" />
