+++ checkstyle/import-control-share.xml
@@ -33,6 +33,7 @@
   <!-- anyone can use public classes -->
   <allow pkg="org.apache.kafka.common" exact-match="true" />
   <allow pkg="org.apache.kafka.common.utils" />
+  <allow pkg="org.apache.kafka.common.annotation" />
 
   <!-- protocol, records and request/response utilities -->
   <allow pkg="org.apache.kafka.common.message" />
