+++ core/src/main/java/kafka/server/share/SharePartition.java
@@ -32,18 +32,18 @@ import org.apache.kafka.common.message.ShareFetchResponseData.AcquiredRecords;
 import org.apache.kafka.common.protocol.Errors;
 import org.apache.kafka.common.record.RecordBatch;
 import org.apache.kafka.common.utils.Time;
-import org.apache.kafka.server.group.share.GroupTopicPartitionData;
-import org.apache.kafka.server.group.share.PartitionAllData;
-import org.apache.kafka.server.group.share.PartitionErrorData;
-import org.apache.kafka.server.group.share.PartitionFactory;
-import org.apache.kafka.server.group.share.PartitionIdLeaderEpochData;
-import org.apache.kafka.server.group.share.PartitionStateBatchData;
-import org.apache.kafka.server.group.share.Persister;
-import org.apache.kafka.server.group.share.PersisterStateBatch;
-import org.apache.kafka.server.group.share.ReadShareGroupStateParameters;
-import org.apache.kafka.server.group.share.TopicData;
-import org.apache.kafka.server.group.share.WriteShareGroupStateParameters;
+import org.apache.kafka.server.share.GroupTopicPartitionData;
+import org.apache.kafka.server.share.PartitionAllData;
+import org.apache.kafka.server.share.PartitionErrorData;
+import org.apache.kafka.server.share.PartitionFactory;
+import org.apache.kafka.server.share.PartitionIdLeaderEpochData;
+import org.apache.kafka.server.share.PartitionStateBatchData;
+import org.apache.kafka.server.share.Persister;
+import org.apache.kafka.server.share.PersisterStateBatch;
+import org.apache.kafka.server.share.ReadShareGroupStateParameters;
 import org.apache.kafka.server.share.ShareAcknowledgementBatch;
+import org.apache.kafka.server.share.TopicData;
+import org.apache.kafka.server.share.WriteShareGroupStateParameters;
 import org.apache.kafka.server.util.timer.Timer;
 import org.apache.kafka.server.util.timer.TimerTask;
 import org.apache.kafka.storage.internals.log.FetchPartitionData;
