+++ core/src/test/java/kafka/server/share/SharePartitionTest.java
@@ -37,14 +37,14 @@ import org.apache.kafka.common.record.MemoryRecordsBuilder;
 import org.apache.kafka.common.record.TimestampType;
 import org.apache.kafka.common.utils.MockTime;
 import org.apache.kafka.common.utils.Time;
-import org.apache.kafka.server.group.share.NoOpShareStatePersister;
-import org.apache.kafka.server.group.share.PartitionFactory;
-import org.apache.kafka.server.group.share.Persister;
-import org.apache.kafka.server.group.share.PersisterStateBatch;
-import org.apache.kafka.server.group.share.ReadShareGroupStateResult;
-import org.apache.kafka.server.group.share.TopicData;
-import org.apache.kafka.server.group.share.WriteShareGroupStateResult;
+import org.apache.kafka.server.share.NoOpShareStatePersister;
+import org.apache.kafka.server.share.PartitionFactory;
+import org.apache.kafka.server.share.Persister;
+import org.apache.kafka.server.share.PersisterStateBatch;
+import org.apache.kafka.server.share.ReadShareGroupStateResult;
 import org.apache.kafka.server.share.ShareAcknowledgementBatch;
+import org.apache.kafka.server.share.TopicData;
+import org.apache.kafka.server.share.WriteShareGroupStateResult;
 import org.apache.kafka.server.util.FutureUtils;
 import org.apache.kafka.server.util.timer.SystemTimer;
 import org.apache.kafka.server.util.timer.SystemTimerReaper;
