+++ share/src/main/java/org/apache/kafka/server/share/DeleteShareGroupStateResult.java
@@ -15,20 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.server.group.share;
+package org.apache.kafka.server.share;
 
-import org.apache.kafka.common.message.WriteShareGroupStateResponseData;
+import org.apache.kafka.common.message.DeleteShareGroupStateResponseData;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 /**
- * This class contains the result from {@link Persister#writeState(WriteShareGroupStateParameters)}.
+ * This class contains the result from {@link Persister#deleteState(DeleteShareGroupStateParameters)}.
  */
-public class WriteShareGroupStateResult implements PersisterResult {
+public class DeleteShareGroupStateResult implements PersisterResult {
     private final List<TopicData<PartitionErrorData>> topicsData;
 
-    private WriteShareGroupStateResult(List<TopicData<PartitionErrorData>> topicsData) {
+    private DeleteShareGroupStateResult(List<TopicData<PartitionErrorData>> topicsData) {
         this.topicsData = topicsData;
     }
 
@@ -36,13 +36,12 @@ public class WriteShareGroupStateResult implements PersisterResult {
         return topicsData;
     }
 
-    public static WriteShareGroupStateResult from(WriteShareGroupStateResponseData data) {
+    public static DeleteShareGroupStateResult from(DeleteShareGroupStateResponseData data) {
         return new Builder()
                 .setTopicsData(data.results().stream()
-                        .map(writeStateResult -> new TopicData<>(writeStateResult.topicId(),
-                                writeStateResult.partitions().stream()
-                                        .map(partitionResult -> PartitionFactory.newPartitionErrorData(partitionResult.partition(), partitionResult.errorCode(), partitionResult.errorMessage()))
-                                        .collect(Collectors.toList())))
+                        .map(deleteStateResult -> new TopicData<>(deleteStateResult.topicId(), deleteStateResult.partitions().stream()
+                                .map(partitionResult -> PartitionFactory.newPartitionErrorData(partitionResult.partition(), partitionResult.errorCode(), partitionResult.errorMessage()))
+                                .collect(Collectors.toList())))
                         .collect(Collectors.toList()))
                 .build();
     }
@@ -55,8 +54,8 @@ public class WriteShareGroupStateResult implements PersisterResult {
             return this;
         }
 
-        public WriteShareGroupStateResult build() {
-            return new WriteShareGroupStateResult(topicsData);
+        public DeleteShareGroupStateResult build() {
+            return new DeleteShareGroupStateResult(topicsData);
         }
     }
 }
