+++ share/src/main/java/org/apache/kafka/server/share/GroupTopicPartitionData.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.server.group.share;
+package org.apache.kafka.server.share;
 
 import java.util.List;
 import java.util.Objects;
