+++ share/src/main/java/org/apache/kafka/server/share/PartitionErrorData.java
@@ -15,11 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.server.group.share;
+package org.apache.kafka.server.share;
 
 /**
- * This interface is implemented by classes used to contain the data for a partition in the
- * interface to {@link Persister}.
+ * This interface is implemented by classes used to contain the data for a partition with error data
+ * in the interface to {@link Persister}.
  */
-public interface PartitionInfoData {
+public interface PartitionErrorData extends PartitionInfoData, PartitionIdData {
+    short errorCode();
+
+    String errorMessage();
 }
