+++ share/src/main/java/org/apache/kafka/server/share/PartitionInfoData.java
@@ -15,12 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.server.group.share;
+package org.apache.kafka.server.share;
 
 /**
- * This interface is implemented by classes used to contain the data for a partition with its partition index
- * in the interface to {@link Persister}.
+ * This interface is implemented by classes used to contain the data for a partition in the
+ * interface to {@link Persister}.
  */
-public interface PartitionIdData extends PartitionInfoData {
-    int partition();
+public interface PartitionInfoData {
 }
