+++ share/src/main/java/org/apache/kafka/server/share/ReadShareGroupStateSummaryResult.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.server.group.share;
+package org.apache.kafka.server.share;
 
 import org.apache.kafka.common.message.ReadShareGroupStateSummaryResponseData;
